requirements.txt
Flask
pyaudio
google-cloud-speech
google-cloud-texttospeech
ibm-watsonx-ai

